requirements.txt
openai>=1.0.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
scipy>=1.10.0
elevenlabs>=0.2.26
python-dotenv>=1.0.0 